double changes

+++ Servers/Filters/vtkVRMLSource.cxx
@@ -217,7 +217,7 @@ void vtkVRMLSource::CopyImporterToOutputs()
         {
         vtkUnsignedCharArray *colorArray = vtkUnsignedCharArray::New();
         unsigned char r, g, b;
-        float* actorColor;
+        double* actorColor;
      
         actorColor = actor->GetProperty()->GetColor();
         r = static_cast<unsigned char>(actorColor[0]*255.0);
